+++ CHANGELOG.md
@@ -39,6 +39,8 @@
 * `transform/clr` component: Added the option to set the `axis` along which to apply CLR. Possible to override
   on workflow level as well (PR #767).
 
+* `annotete/scanvi` component: Added a component to annotate cells using scANVI (PR #833).
+
 ## MINOR CHANGES
 
 * Bump scvelo to `0.3.2` (PR #828).